Create PSDetect.py
In this program, I built a port scanner detector, PSDetect. PSDetect uses pycapy1 to listen to incoming connections, and report the presence of a scanner if a single machine (identified by a single source IP address) attempted to connect to 30 or more ports within a 5 second window. PSDetect is able to detect when PortScan is used.